requirements.txt
autopep8==1.5.7
cycler==0.10.0
kiwisolver==1.3.1
numpy==1.21.1
opencv-python==4.5.3.56
Pillow==8.3.1
pycodestyle==2.7.0
pyparsing==2.4.7
pytesseract==0.3.8
python-dateutil==2.8.2
selenium==3.141.0
six==1.16.0
toml==0.10.2
urllib3==1.26.6
